Address issues discussed a few weeks back
This covers the issues windows was having with CMake we discussed w few weeks back. The project is still locked to C++17 due to the past issues we discussed. Even though 17 is a pretty decent minimum I still think I'd be worth further investigating this issue. For now this is a bandaid to make SFML work for windows users without needing to figure it out on there own.